+++ src/main/java/com/room414/racingbets/dal/abstraction/dao/RacecourseDao.java
@@ -11,11 +11,11 @@ import java.util.List;
  * @author Alexander Melashchenko
  * @version 1.0 27 Feb 2017
  */
-public interface RacecourseDao extends CrudDao<Integer, Racecourse> {
+public interface RacecourseDao extends CrudDao<Long, Racecourse> {
     /**
      * @return List of Racecourses which names starts with namePart or empty list if no found.
      */
-    List<Racecourse> findByNamePart(String namePart, int offset, int limit);
+    List<Racecourse> findByNamePart(String namePart, long offset, long limit);
 
     /**
      * @return count of Racecourses which names starts with namePart
